Moved Sidebar annotations => org.vaadin.spring.sidebar.annotation
