+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/ExecutionOperation3.java
@@ -19,8 +19,8 @@ import com.vaadin.ui.Notification;
 
 import org.springframework.stereotype.Component;
 import org.vaadin.spring.annotation.VaadinUIScope;
-import org.vaadin.spring.sidebar.SideBarItem;
-import org.vaadin.spring.sidebar.ThemeIcon;
+import org.vaadin.spring.sidebar.annotation.SideBarItem;
+import org.vaadin.spring.sidebar.annotation.ThemeIcon;
 
 import java.io.Serializable;
