+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/PlanningOperation.java
@@ -20,8 +20,8 @@ import com.vaadin.ui.Notification;
 
 import org.springframework.stereotype.Component;
 import org.vaadin.spring.annotation.VaadinUIScope;
-import org.vaadin.spring.sidebar.FontAwesomeIcon;
-import org.vaadin.spring.sidebar.SideBarItem;
+import org.vaadin.spring.sidebar.annotation.FontAwesomeIcon;
+import org.vaadin.spring.sidebar.annotation.SideBarItem;
 
 import java.io.Serializable;
