+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/PlanningView2.java
@@ -23,8 +23,8 @@ import com.vaadin.ui.VerticalLayout;
 
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.annotation.VaadinView;
-import org.vaadin.spring.sidebar.FontAwesomeIcon;
-import org.vaadin.spring.sidebar.SideBarItem;
+import org.vaadin.spring.sidebar.annotation.FontAwesomeIcon;
+import org.vaadin.spring.sidebar.annotation.SideBarItem;
 
 /**
  * Example view that shows up under the Planning section in the side bar.
