+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/ReportingView1.java
@@ -22,8 +22,8 @@ import com.vaadin.ui.VerticalLayout;
 
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.annotation.VaadinView;
-import org.vaadin.spring.sidebar.SideBarItem;
-import org.vaadin.spring.sidebar.ThemeIcon;
+import org.vaadin.spring.sidebar.annotation.SideBarItem;
+import org.vaadin.spring.sidebar.annotation.ThemeIcon;
 
 /**
  * Example view that shows up under the Reporting section in the side bar.
