+++ spring-vaadin-sidebar/src/main/java/org/vaadin/spring/sidebar/annotation/FontAwesomeIcon.java
@@ -13,26 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.spring.sidebar;
+package org.vaadin.spring.sidebar.annotation;
+
+import com.vaadin.server.FontAwesome;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.vaadin.spring.sidebar.FontAwesomeIconProvider;
+
 /**
- * This annotation is placed next to {@link org.vaadin.spring.sidebar.SideBarItem}, instructing the side bar
- * to use a {@link com.vaadin.server.ThemeResource} icon for the item.
+ * This annotation is placed next to {@link org.vaadin.spring.sidebar.annotation.SideBarItem}, instructing the side bar
+ * to use a {@link com.vaadin.server.FontAwesome} icon for the item.
  *
  * @author Petter Holmström (petter@vaadin.com)
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@SideBarItemIcon(ThemeIconProvider.class)
-public @interface ThemeIcon {
+@SideBarItemIcon(FontAwesomeIconProvider.class)
+public @interface FontAwesomeIcon {
 
     /**
-     * The theme resource ID of the item icon.
+     * The item icon.
      */
-    String value();
+    FontAwesome value();
 }
