+++ spring-vaadin-sidebar/src/main/java/org/vaadin/spring/sidebar/annotation/SideBarSections.java
@@ -13,22 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.spring.sidebar;
+package org.vaadin.spring.sidebar.annotation;
 
-import org.springframework.context.annotation.Import;
-import org.vaadin.spring.sidebar.config.SideBarConfiguration;
-
-import java.lang.annotation.*;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * Add this annotation to your application configuration to enable the {@link org.vaadin.spring.sidebar.SideBar}
- * component. After that, just inject the side bar into your UIs.
+ * This annotation is used to declare multiple {@link org.vaadin.spring.sidebar.annotation.SideBarSection}s on the same
+ * bean. Please refer to the JavaDocs for {@link org.vaadin.spring.sidebar.annotation.SideBarSection} for more information.
  *
  * @author Petter Holmström (petter@vaadin.com)
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Import(SideBarConfiguration.class)
-public @interface EnableSideBar {
+public @interface SideBarSections {
+
+    SideBarSection[] value();
 }
